+++ src/main/adiru-nosocket.cpp
